Refactor zeroLive to utilize ZeroStoreLive and update mutator imports for consistency. Remove deprecated effectMutatorConverter and effectTransaction files to streamline the codebase.
